+++ docker/cs-web-server/src/client/main.ts
@@ -2,9 +2,10 @@ import {loadAsync} from 'jszip'
 import filesystemURL from 'xash3d-fwgs/filesystem_stdio.wasm?url'
 import xashURL from 'xash3d-fwgs/xash.wasm?url'
 import menuURL from 'cs16-client/cl_dll/menu_emscripten_wasm32.wasm?url'
-import clientURL from 'cs16-client/cl_dll/client_emscripten_wasm32.wasm?url'
+import clientURL from 'cs16-client/cl_dlls/client_emscripten_wasm32.wasm?url'
 import serverURL from 'cs16-client/dlls/cs_emscripten_wasm32.so?url'
-import gles3URL from 'xash3d-fwgs/libref_gles3compat.wasm?url'
+import gl4esURL from 'xash3d-fwgs/libref_webgl2.wasm?url'
+import extrasURL from 'cs16-client/extras.pk3?url'
 import {Xash3DWebRTC} from "./webrtc";
 
 let usernamePromiseResolve: (name: string) => void
@@ -15,9 +16,7 @@ const usernamePromise = new Promise<string>(resolve => {
 async function main() {
     const x = new Xash3DWebRTC({
         canvas: document.getElementById('canvas') as HTMLCanvasElement,
-        module: {
-            arguments: ['-windowed', '-game', 'cstrike'],
-        },
+        arguments: ['-windowed', '-game', 'cstrike'],
         libraries: {
             filesystem: filesystemURL,
             xash: xashURL,
@@ -25,19 +24,24 @@ async function main() {
             server: serverURL,
             client: clientURL,
             render: {
-                gles3compat: gles3URL,
+                gl4es: gl4esURL,
             }
         },
+        dynamicLibraries: ['dlls/cs_emscripten_wasm32.so', '/rwdir/filesystem_stdio.wasm'],
         filesMap: {
             'dlls/cs_emscripten_wasm32.so': serverURL,
-            '/rwdir/filesystem_stdio.so': filesystemURL,
+            '/rwdir/filesystem_stdio.wasm': filesystemURL,
         },
     });
 
-    const [zip] = await Promise.all([
+    const [zip, extras] = await Promise.all([
         (async () => {
             const res = await fetch('valve.zip')
             return await loadAsync(await res.arrayBuffer());
+        })(),
+        (async () => {
+            const res = await fetch(extrasURL)
+            return await res.arrayBuffer();
         })(),
         x.init(),
     ])
@@ -52,6 +56,9 @@ async function main() {
         x.em.FS.writeFile(path, await file.async("uint8array"));
     }))
 
+    x.em.FS.writeFile('/rodir/cstrike/extras.pk3', new Uint8Array(extras))
+    const isMobile = !window.matchMedia('(hover: hover)').matches
+
     x.em.FS.chdir('/rodir')
 
     document.getElementById('logo')!.style.animationName = 'pulsate-end'
@@ -62,7 +69,7 @@ async function main() {
     const username = await usernamePromise
     x.main()
     x.Cmd_ExecuteString('_vgui_menus 0')
-    if (!window.matchMedia('(hover: hover)').matches) {
+    if (isMobile) {
         x.Cmd_ExecuteString('touch_enable 1')
     }
     x.Cmd_ExecuteString(`name "${username}"`)